To emulate missed parts of DSObject - optional file_path removing #1241

+++ src/sugar/datastore/datastore.py
@@ -170,6 +170,7 @@ class RawObject(object):
     def get_file_path(self, fetch=True):
         # we have to create symlink since its a common practice
         # to create hardlinks to jobject files
+        # and w/o this, it wouldn't work since we have file from mounted device
         if self._file_path is None:
             self._file_path = tempfile.mktemp(
                     prefix='rawobject',
@@ -185,7 +186,8 @@ class RawObject(object):
             return
         self._destroyed = True
         if self._file_path is not None:
-            os.remove(self._file_path)
+            if os.path.exists(self._file_path):
+                os.remove(self._file_path)
             self._file_path = None
 
     def __del__(self):
